apps/digit_recognizer_web/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy>=1.26.0
pillow>=10.0.0
